feat: rpc node management for optimism
