+++ frontend/app/src/components/helper/ListItem.vue
@@ -119,10 +119,6 @@ const click = () => emit('click');
   text-overflow: ellipsis;
   overflow: hidden;
   white-space: nowrap;
-
-  @media (min-width: 700px) and (max-width: 1200px) {
-    width: 100px;
-  }
 }
 
 .title {
@@ -130,10 +126,6 @@ const click = () => emit('click');
   overflow: hidden;
   white-space: nowrap;
   font-weight: 500;
-
-  @media (min-width: 700px) and (max-width: 1200px) {
-    width: 100px;
-  }
 }
 
 .subtitle {
@@ -141,9 +133,5 @@ const click = () => emit('click');
   text-overflow: ellipsis;
   overflow: hidden;
   white-space: nowrap;
-
-  @media (min-width: 700px) and (max-width: 1200px) {
-    width: 100px;
-  }
 }
 </style>
