+++ frontend/app/src/components/settings/general/rpc/EthRpcSetting.vue
@@ -1,0 +1,9 @@
+<script lang="ts" setup>
+import { Blockchain } from '@rotki/common/lib/blockchain';
+
+const chain = Blockchain.ETH;
+</script>
+
+<template>
+  <evm-rpc-node-manager :chain="chain" />
+</template>
