+++ frontend/app/src/components/settings/general/rpc/OptimismRpcSetting.vue
@@ -1,0 +1,9 @@
+<script lang="ts" setup>
+import { Blockchain } from '@rotki/common/lib/blockchain';
+
+const chain = Blockchain.OPTIMISM;
+</script>
+
+<template>
+  <evm-rpc-node-manager :chain="chain" />
+</template>
