+++ frontend/app/src/composables/api/settings/evm-nodes-api.ts
@@ -1,0 +1,54 @@
+import { type ActionResult } from '@rotki/common/lib/data';
+import { Blockchain } from '@rotki/common/lib/blockchain';
+import { axiosSnakeCaseTransformer } from '@/services/axios-tranformers';
+import { api } from '@/services/rotkehlchen-api';
+import { handleResponse, validStatus } from '@/services/utils';
+import { type EvmRpcNode, EvmRpcNodeList } from '@/types/settings';
+
+export const useEvmNodesApi = (chain: Blockchain = Blockchain.ETH) => {
+  const url = `/blockchains/${chain}/nodes`;
+
+  const fetchEvmNodes = async (): Promise<EvmRpcNodeList> => {
+    const response = await api.instance.get<ActionResult<EvmRpcNodeList>>(url);
+    return EvmRpcNodeList.parse(handleResponse(response));
+  };
+
+  const addEvmNode = async (
+    node: Omit<EvmRpcNode, 'identifier'>
+  ): Promise<boolean> => {
+    const response = await api.instance.put<ActionResult<boolean>>(
+      url,
+      axiosSnakeCaseTransformer(node),
+      {
+        validateStatus: validStatus
+      }
+    );
+    return handleResponse(response);
+  };
+
+  const editEvmNode = async (node: EvmRpcNode): Promise<boolean> => {
+    const response = await api.instance.patch<ActionResult<boolean>>(
+      url,
+      axiosSnakeCaseTransformer(node),
+      {
+        validateStatus: validStatus
+      }
+    );
+    return handleResponse(response);
+  };
+
+  const deleteEvmNode = async (identifier: number): Promise<boolean> => {
+    const response = await api.instance.delete<ActionResult<boolean>>(url, {
+      data: axiosSnakeCaseTransformer({ identifier }),
+      validateStatus: validStatus
+    });
+    return handleResponse(response);
+  };
+
+  return {
+    fetchEvmNodes,
+    addEvmNode,
+    editEvmNode,
+    deleteEvmNode
+  };
+};
