+++ frontend/app/src/types/settings.ts
@@ -1,8 +1,9 @@
 import { z } from 'zod';
+import { type Blockchain } from '@rotki/common/lib/blockchain';
 
-const EthereumRpcNode = z.object({
+const EvmRpcNode = z.object({
   identifier: z.number(),
-  name: z.string().nonempty(),
+  name: z.string().min(1),
   endpoint: z.string(),
   owned: z.boolean(),
   weight: z.preprocess(
@@ -10,20 +11,20 @@ const EthereumRpcNode = z.object({
     z.number().nonnegative().max(100)
   ),
   active: z.boolean(),
-  blockchain: z.string().nonempty()
+  blockchain: z.string().min(1)
 });
 
-export type EthereumRpcNode = z.infer<typeof EthereumRpcNode>;
+export type EvmRpcNode = z.infer<typeof EvmRpcNode>;
 
-export const EthereumRpcNodeList = z.array(EthereumRpcNode);
-export type EthereumRpcNodeList = z.infer<typeof EthereumRpcNodeList>;
+export const EvmRpcNodeList = z.array(EvmRpcNode);
+export type EvmRpcNodeList = z.infer<typeof EvmRpcNodeList>;
 
-export const getPlaceholderNode = (): EthereumRpcNode => ({
+export const getPlaceholderNode = (chain: Blockchain): EvmRpcNode => ({
   identifier: -1,
   name: '',
   endpoint: '',
   weight: 0,
   active: true,
   owned: true,
-  blockchain: 'ETH'
+  blockchain: chain
 });
